rutil: Data: use of fixed-size integer types for methods that work with fixed buffer sizes
git-svn-id: https://svn.resiprocate.org/rep/resiprocate/main@10433 ddefafc4-47db-0310-ae44-fa13212b10f2

+++ resip/stack/TransactionState.cxx
@@ -43,7 +43,7 @@ using namespace resip;
 
 #define RESIPROCATE_SUBSYSTEM Subsystem::TRANSACTION
 
-unsigned long TransactionState::StatelessIdCounter = 0;
+UInt32 TransactionState::StatelessIdCounter = 0;
 
 TransactionState::TransactionState(TransactionController& controller, Machine m, 
                                    State s, const Data& id, MethodTypes method, const Data& methodText, TransactionUser* tu) :
